requirements.txt
absl-py==2.1.0
accelerate==1.2.1
adabelief-pytorch==0.2.1
addict==2.4.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
albucore
albumentations
annotated-types
async-timeout==5.0.1
attrs==24.2.0
basicsr==1.4.2
betacal==1.1.0
Bottleneck
box==0.1.5
calflops==0.3.2
causal-conv1d==1.5.0.post8
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cloudpickle==3.1.1
colorama==0.4.6
Columnar==1.3.1
contourpy==1.3.0
cvxopt==1.3.2
cycler==0.12.1
diffusers==0.31.0
docopt==0.6.2
einops==0.8.0
et_xmlfile==2.0.0
eval_type_backport
exceptiongroup==1.2.2
executing==0.8.2
facexlib==0.3.0
faster-coco-eval==1.6.5
filelock==3.16.1
filterpy==1.4.5
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.10.0
ftfy==6.3.1
future==1.0.0
gfpgan==1.3.8
grad-cam==1.5.5
grpcio==1.68.1
h5py==3.12.1
huggingface-hub==0.27.0
idna==3.10
imagecodecs
imageio
imbalanced-learn==0.12.4
imblearn==0.0
importlib_metadata
importlib_resources==6.4.5
iniconfig==2.0.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
kornia==0.7.3
kornia_rs==0.1.7
lazy_loader
lightning-utilities==0.11.9
llvmlite==0.43.0
lmdb==1.5.1
loguru==0.7.2
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.2
MedPy==0.5.2
ml_collections==1.0.0
ml_insights==1.0.4
mlinsights==0.5.1
mpmath==1.3.0
multidict==6.1.0
narwhals==1.26.0
networkx
nibabel==5.3.2
ninja==1.11.1.3
numba==0.60.0
numexpr
numpy==1.24.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
onnx==1.17.0
open_clip_torch @ git+https://github.com/mlfoundations/open_clip.git@63fbfd857c83af619e6a0a8344635ebb4a151a96
opencv-python==4.10.0
opencv-python-headless==4.10.0
openpyxl==3.1.5
packaging
pandas==2.2.3
pillow
platformdirs==4.3.6
plotly==6.0.0
pluggy==1.5.0
propcache==0.2.1
protobuf==5.28.3
psutil==6.1.0
py-cpuinfo==9.0.0
pyaml==24.9.0
pycocotools==2.0.8
pydantic
pydantic_core
pydicom==3.0.1
pykwalify==1.8.0
pyparsing==3.2.0
pyradiomics==3.0.1
pysqlite3-binary==0.5.4
pytest==8.3.4
python-dateutil==2.8.2
pytorch-lightning==2.5.0.post0
pytz
PyWavelets==1.7.0
PyYAML
realesrgan==0.3.0
regex==2024.11.6
requests==2.32.3
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
safetensors==0.4.5
scikit-image
scikit-learn==1.5.2
scikit-optimize==0.10.2
scipy==1.14.1
seaborn==0.13.2
SimpleITK==2.4.0
six
splinecalib==0.0.9
submitit==1.5.2
sympy==1.13.1
tabulate==0.9.0
tb-nightly==2.19.0a20241216
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.5.0
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile
timeago==1.0.14
timm==1.0.12
tokenizers==0.21.0
tomli
toolz==1.0.0
torch==2.5.1
torchaudio==2.5.1
torchmetrics==1.6.1
torchvision==0.20.1
tqdm==4.66.6
transformers==4.47.0
triton==3.1.0
ttach==0.0.3
typing_extensions
tzdata
ultralytics==8.3.78
ultralytics-thop==2.0.14
urllib3==2.2.3
vit-pytorch==1.8.7
wcwidth==0.2.13
Werkzeug==3.1.3
xlrd==2.0.1
xlwt==1.3.0
yacs==0.1.8
yapf==0.43.0
yarl==1.18.3
zipp
